feat: removed middle step and submit to discord directly in server action
